avances en el grid
