+++ Frontend/src/Components/itemVerProductos.jsx
@@ -1,27 +1,38 @@
 import React from "react";
-import itemGrid from "../styles/itemGrid.css";
+import "../styles/itemGrid.css";
 import "../styles/VerProductos.css"
+
+const handleDeleteItem = (itemId, setItems) => {
+  var requestOptions = {
+    method: 'POST',
+    body: JSON.stringify({
+      itemId,
+    }),
+    redirect: 'follow'
+  };
+
+  fetch("http://127.0.0.1:8000/deleteItem", requestOptions)
+    .then(response => response.json())
+    .then(result => {
+      console.log(result.succes),
+      setItems([])
+    })
+    .catch(error => console.log('Error en handleItemsUser', error));
+}
+
+
 
 const ItemVerProductos = ({
   itemName,
   itemLocation,
   itemPrice,
   itemPublishDate,
-  itemImage,
-  itemId,
-  setIdSelectedProduct,
-  setSelectedProduct
+  itemImage
 }) => {
 
-  const handleDelete = () => {
-    
-  }
 
   return (
-    <div className="container_item" onClick={()=> {
-      setIdSelectedProduct(itemId),
-      setTimeout(setSelectedProduct(true), 100)
-    }}>
+    <div className="container_item">
       <figure className="fotositem_portada">
         <img className="productImage" src={itemImage}/>
       </figure>
@@ -32,9 +43,8 @@ const ItemVerProductos = ({
       <h3 className="product_name"> {itemName}</h3>
       <div className="contenidositem_tiemporating">
         <h6 className="hora_publicacion">{itemPublishDate} </h6>
-        {/* <h3 className="hora_publicacion">Publicado el {itemPublishDate[8]}{itemPublishDate[9]}/{itemPublishDate[5]}{itemPublishDate[6]}/{itemPublishDate[0]}{itemPublishDate[1]}{itemPublishDate[2]}{itemPublishDate[3]}</h3> */}
       </div>
-      <button className="btn_delete" onClick={() => handleDetele()}>x</button>
+      <button className="btn_delete" onClick={() => handleDeleteItem()}>x</button>
     </div>
   );
 };
