+++ Frontend/src/styles/VerProductos.css
@@ -1,4 +1,10 @@
 .btn_delete{
     background-color: red;
-    border-radius: 6px;
+    border-radius: 50%;
+    width: 50px;
+    height: 50px;
+    border:none;
+    font-size: 20px;
+
+    
 }
