+++ Frontend/src/styles/login.css
@@ -95,9 +95,7 @@ input {
   flex-direction: row;
 }
 
-.main-content-login {
-  margin-top: 90px;
-}
+
 
 .btnLogin {
   padding: 15px 20px;
