+++ Frontend/src/styles/navbar.css
@@ -12,6 +12,7 @@
   display: flex;
   flex-direction: row;
   justify-content: space-around;
+  margin-bottom: 3%;
 }
 
 .navbar div {
